Osc config load: rework to avoid cross dependencies
Load all the config in a map, then iterate each nodes that have a value
and execute them with the config value if it exists.
Do this in a loop until no new nodes need configuration.

This handle the case when nodes are created by specific values or an OSC
variable (especially type and endpoints).